+++ lib/src/middleware/cors.dart
@@ -1,12 +1,14 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:alfred/alfred.dart';
+
 /// CORS Middlware.
 ///
 /// Has some sensible defaults. You probably want to change the origin
 FutureOr Function(HttpRequest, HttpResponse) cors(
     {int age = 86400,
-    String headers = 'Content-Type',
+    String headers = '*',
     String methods = 'POST, GET, OPTIONS, PUT, PATCH',
     String origin = '*'}) {
   return (HttpRequest req, HttpResponse res) {
@@ -14,5 +16,8 @@ FutureOr Function(HttpRequest, HttpResponse) cors(
     res.headers.set('Access-Control-Allow-Methods', methods);
     res.headers.set('Access-Control-Allow-Headers', headers);
     res.headers.set('Access-Control-Max-Age', age);
+    if (req.method == 'OPTIONS') {
+      res.close();
+    }
   };
 }