+++ lib/src/type_handlers/serializable_type_handler.dart
@@ -26,4 +26,6 @@ TypeHandler<dynamic> get serializableTypeHandler =>
           rethrow;
         }
       }
+
+      return false;
     });